refactor(views): migrate albums to per-folder layout (commit 4/7)
Split src/views/albums.rs (1469 lines) into src/views/albums/{mod,update,view}.rs
following the genres/playlists template:

- mod.rs: types (Page, Message, Action, ViewData), trait impls
  (HasCommonAction via macro, ViewPage explicit), Default + new(),
  test for column-visibility default
- update.rs: impl AlbumsPage { fn update } + tests for
  center-on-playing and toggle-column
- view.rs: impl AlbumsPage { fn view, fn render_album_row,
  fn render_track_row } + the per-mode column helpers
  (albums_stars_visible / albums_plays_visible) + the dynamic-slot
  value resolver (get_extra_column_value) + helper tests

Macro consumers:
- define_view_columns! for AlbumsColumn / Visibility (7 columns).
- impl_has_common_action!(AlbumsAction { NavigateAndExpandArtist,
  NavigateAndExpandGenre }).

The album view kept its overlay text rendering (stars-on-Rating-sort,
plays-on-MostPlayed-sort, dynamic Date/Year/Duration/Genre slot)
identical — no behavioral change.

cargo build / clippy --all-targets -D warnings / nightly fmt --check /
test --workspace all clean (662 + 368 + 1 doctest).
